regexp: port RE2 DFA construction code
some hacking on my flights to and from gophercon
+ some more hacking during winter break

DO NOT REVIEW

(you've been warned)

This is a dumb port of the DFA construction code in RE2
(that's most of the stuff rooted ad BuildAllStates()).
It's very messy: full of garbage and notes to myself.

None of the Search functions have been ported, but for simple regexps,
you can call buildAllStates, and then use search to hop between the
states and determine whether there's a match.

But there's a lot missing and broken. Among other things:
- None of the search methods have been ported
- The input string can't contain any runes > 0xFF
- Flags, params, and empty ops don't work
- I've done some things -- especially state allocation, and state cache lookups
  -- in a very dumb, inefficient way
- Atomic loads and stores

Change-Id: If60bb9c8586bc02aefd5251dff0acaa46cd6f7e4